# Conflicts:
#	src/main/java/com/SmartBank/dto/response/TransactionResponse.java
#	src/main/java/com/SmartBank/repository/AccountRepository.java
#	src/main/java/com/SmartBank/service/CustomerService.java

+++ src/main/java/com/SmartBank/service/CustomerService.java
@@ -53,7 +53,7 @@ public class CustomerService {
                 .collect(Collectors.toList());
     }
 
-    public CustomerResponse getById(Long id) {
+    public CustomerResponse getById(Integer id) {
         Customer customer = repository.findById(id).orElse(null);
         if(customer == null) {
             throw new ResourceNotFoundException("Customer not found");
@@ -61,7 +61,7 @@ public class CustomerService {
         return customerMapper.toResponse(customer);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Integer id) {
         Customer customer = repository.findById(id).orElse(null);
         if(customer == null) {
             throw new ResourceNotFoundException("Customer not found");
@@ -70,7 +70,7 @@ public class CustomerService {
         repository.deleteById(id);
     }
 
-    public CustomerResponse update(Long id, CustomerRequest request) {
+    public CustomerResponse update(Integer id, CustomerRequest request) {
         Customer customer = repository.findById(id).orElse(null);
         if(customer == null) {
             throw new ResourceNotFoundException("Customer not found");
@@ -85,7 +85,7 @@ public class CustomerService {
         return customerMapper.toResponse(customer);
     }
 
-    public @Nullable List<AccountResponse> getAccountsByID(Long id) {
+    public @Nullable List<AccountResponse> getAccountsByID(Integer id) {
         Customer customer = repository.findById(id).orElse(null);
         if(customer == null) {
             throw new ResourceNotFoundException("Customer not found");
